Update filter logic and UI enhancements in product list
Revised filter operations in product list to manage search data and updated the user interface for the product filter components. This involves an important change in the search criteria in product services. In addition, the default visibility of the product filters has been set to true and icon images have been relocated to a new path.