Add appearance order to menu and category items
The code was modified to include an appearance order number to the category and menu interfaces. This will allow for customizable order of display in the navigation menu list. Also, made some adjustments in the display properties of various components and some hostname entries in the next.config.js file.