Add missing product image in image list
A new element has been added to the image list in product-detail component. This additional image has a "missing product image" placeholder which appears when the mouse hovers over it. This results in a better user experience, as it prevents empty spaces from appearing in case of missing product images.